qb_app/requirements.txt
Flask
requests
pyodbc
python-dotenv
PyJWT
bcrypt
Flask-Cors
